requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
deepmerge==1.1.1
Faker==22.1.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.0.11
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
json_diff==1.5.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
locust==2.23.1
MarkupSafe==2.1.5
msgpack==1.0.7
mysql-connector-python==8.2.0
outcome==1.3.0.post0
packaging==23.2
playwright==1.42.0
pluggy==1.4.0
protobuf==4.21.12
psutil==5.9.8
pycparser==2.21
pyee==11.0.1
PySocks==1.7.1
pytest==8.0.0
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.4
pywin32==306
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
roundrobin==0.0.4
rpds-py==0.17.1
selenium==4.16.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.2
